[contacts] Major refactoring of contacts part of the app.
• all contacts are now loaded as soon as app starts for the first time,
  and into a separate database, placed in app group containter (which
  means it would be accessible to app extensions too).
• contacts are synced with the server using new APIs.
• there's a new SyncManager class that is responsible for all contacts
  sync logic and scheduling.

Bonus: Messages view (for now) will show all contacts that are using
HalloApp. If contact name is in black it means that you're connected.
If contact name is in grey it means that contact doesn't have your phone
number saved and you can't see each others posts.

There's still a good amount of TODOs and things to optimize: to be
addressed in the future.