Prepare to support app extensions.
There's now an internal framework named "Core" in the project. This
framework is designed to contain only the code that must be shared
between app and it's extensions.

Also added to the project but not included with app installation
is our very first app extension: Notification Service Extension.
For now this extension does nothing, but in the future it will decode
protobuf content server will send in APNS notification payload and
populate notification presented to the user with real data.