Show local notifications for replies to your comments.
Server won't send regular push notifications for replies to your comments
because it doesn't know the author of the original comment. Instead,
server will now send silent pushes for all comments to all posts visible
to the user. Once app connects in the background it'll have a chance
to present notifications for all comment replies.