Fix bug where some feed media couldn't be downloaded.
Path to a media file was constructed using first two letters of
feed post id like this: "Media/A/B/AB<rest of the id>.jpg". Feed post ids
coming from Android devices are lowercase and iOS is uppercase. As a result,
media from Android devices couldn't be saved if there already was
a directory with the same letter, but uppercase.

The fix is to convert all letters to uppercase when constructing path.