Show "loading" in Comments screen opened for post that app doesn't have (yet).
This situation is possible when user taps on a comment notification
and app opens Comments view automatically.

It is unlikely to receive a notification for a comment on a post
that local storage does not have yet (it must be a someone else's post
that current user has not interacted with yet), but nevertheless the
support for this scenario is there.