Incoming chat messages increase app icon badge number.
Current badge number is stored in user defaults and is accessible to
main app and share extensions.

Notification service extension increments number when processing
a notification for a chat message and then updates notification with
the new number, which is shown on app icon.

Opening the main app resets (removes) badge number.