Fix device hang when displaying NUX
Adding the overlay container to the tab bar controller's view was reliably freezing my phone.

I wasn't able to reproduce it in the simulator, but on device UIKit was getting stuck in a -layoutSubviews loop.

This fix isn't ideal (there is an extra gap at the bottom where the tab bar was, and the navbar brightens). But works for now.