Fix comments could be cut off in some cases (#140).
The issue was caused by showing / hiding activity indicator in a comment
table view cell without invoking any of UITableView's methods: it is an
programmer error to refresh UITableViewCell in place if such refresh
causes change in cell's intrinsic content size.

The fix is to show activity indicator (while comment is being posted)
instead of comment timestamp and Reply button.