Fix the keyboard bug in #471
The issue stems from the bottom padding of the outer VStack,
which in turn makes the inner card top positioned ABOVE
the start of the scroll view if there is not enough vertical space.
This triggers the overscroll logic every time and so the keyboard
is always closed and cannot be accessed.
Set the outer VStack height directly using frame to avoid the
complications from specifying a padding.