Update contact sorting
We want new threads to start at the top, so let's set lastMsgTimestamp to sort it as if there was a new message.

We don't want to lock unvisited threads at the top, so don't sort based on isNew.

Now that we're setting lastMsgTimestamp on new threads, we need to make sure there was a lastMsgId before showing the last message time.