Remove FeedDataItem (#1122)
We're seeing some issues synchronizing the list of FeedDataItems with the UI (empty posts, etc.). The items are also problematic because they hold strong references to images we'd like to be able to unload to reduce memory usage. For now let's move the media into a central cache and later we can update the media types to support unloading.