[Chat] Band-aid fix for crashes when there's unknown media types
Remove force unwrapping for some managedObject properties and default them to none
or image.  Ideally, it should be none as image introduces unsightly UI rendering in
the chats screen due to it expecting an image instead of an unknown type.
However, this is still better than the app crashing and this issue (crash) should
not occur unless the user downgrades the app.