Pop the stack of views when selecting any new tab
This makes sense for the groups tab and the chats tab.  I've generally
found it annoying that the settings tab stays wherever you last visited
so this also resets to the root view for settings, but I see other apps
also maintain the navigation position on the settings tab.

Fixes 735