Replaces fullscreen navigation bar with custom navigation
Since iOS 11 the native buttons in navigation bar are 8pts from the edges, but when using custom buttons it is 16pts.

In fullscreen when a transition is happening the back button below might be visible, for example between a group feed. Due to the different positions the back buttons of the two screens are visibly not aligned.

We've been trying to align those buttons, but the end result is that to accomplish this the button must be partially outside its parent view in the navigation bar. This has the undesirable effect of reducing the tapable area.

So in this PR I've removed the navigation bar and put a custom view instead so both the buttons are well positioned and fully tapable.