Fixed volume button triggering camera shutter
There already is an implementation for this feature in CameraController. The issue is that it observes an undocumented notification for volume changes that has been renamed in iOS 15.

It's not a clean solution, but it seems like the only way to achieve this functionality. I placed a detailed comment in the code that breaks it down.