Resolves #1849. Starts media processing early (#1952)
Moves media processing much earlier, as soon as the composer is open and the media is ready. As a result, the whole sharing process should feel much faster to the user and it will also require less background time.

An added benefit is that, unlike before, in the share extension a media is processed only once even when multiple destinations are selected. It is still uploaded multiple times but I am going to look into how to improve that in a future PR.