Fixes interactive back transition when keyboard is open
When the keyboard is open in chat or comments and you open in fullscreen,
then when you return using an interactive transition, by swiping down,
the back transition might have a wrong position. This happens because at
the moment when the position is taken the keyboard is now shown, and then
at the moment the transition is happening the keyboard is shown.